added customer validation#

+++ src/main/java/com/ludowica/salesapi/controllers/CustomerController.java
@@ -1,8 +1,11 @@
 package com.ludowica.salesapi.controllers;
 
 import com.ludowica.salesapi.models.Customer;
+import com.ludowica.salesapi.models.Enquiry;
 import com.ludowica.salesapi.repository.CustomerRepo;
+import com.ludowica.salesapi.repository.EnquiryRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,28 +18,39 @@ public class CustomerController {
     @Autowired
     CustomerRepo customerRepo;
 
+    @Autowired
+    EnquiryRepo enquiryRepo;
+
+
     @GetMapping
-    public List<Customer> getAllCustomers(){
+    public List<Customer> getAllCustomers() {
         return customerRepo.findAll();
     }
 
     @GetMapping("/{id}")
-    public Optional<Customer> getCustomer(@PathVariable int id){
+    public Optional<Customer> getCustomer(@PathVariable int id) {
         return customerRepo.findById(id);
     }
 
     @PostMapping
-    public Customer addCustomer(@RequestBody Customer customer){
+    public Customer addCustomer(@RequestBody Customer customer) {
         return customerRepo.save(customer);
     }
 
     @PutMapping
-    public Customer updateCustomer(@RequestBody Customer customer){
+    public Customer updateCustomer(@RequestBody Customer customer) {
         return customerRepo.save(customer);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCustomer(@PathVariable int id){
+    public void deleteCustomer(@PathVariable int id) {
+
+        Optional<Enquiry> enquiry = enquiryRepo.findByCustomerId(id);
+
+        if(enquiry.isPresent()){
+            throw new ResourceNotFoundException("Customer Has an Enquiry");
+        }
+
         customerRepo.deleteById(id);
     }
